# docs/openapi.yaml
openapi: 3.0.3
info:
  title: APEX Router API
  description: |
    APEX is a pre-alpha autonomous agent platform. 
    
    ## Authentication
    All API requests require HMAC signature authentication:
    - `X-APEX-Signature`: HMAC-SHA256 signature
    - `X-APEX-Timestamp`: Unix timestamp (within 5 minutes)
    
    Set `APEX_AUTH_DISABLED=1` for local development.
    
    ## Base URL
    `http://localhost:3000`
  version: 1.3.0
  contact:
    name: APEX Team
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:3001
    description: Gateway (proxied)
...
paths:
...
  /api/v1/tasks:
    get:
      summary: List tasks
      description: List all tasks with optional filtering
      tags: [Tasks]
      parameters:
        - name: project
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create task
      description: Create a new task. Auto-tiers based on content classification.
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /api/v1/tasks/{task_id}:
    get:
      summary: Get task
      description: Get task status and result
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
    put:
      summary: Update task
      description: Update task fields (project, priority, category, status)
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
  /api/v1/tasks/{task_id}/cancel:
    post:
      summary: Cancel task
      description: Cancel a running or pending task
      tags: [Tasks]
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
...
components:
  schemas:
    Task:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        content:
          type: string
        output:
          type: string
        error:
          type: string
        project:
          type: string
        priority:
          type: string
        category:
          type: string
        created_at:
          type: string
          format: date-time
    CreateTaskRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        channel:
          type: string
        thread_id:
          type: string
        author:
          type: string
        max_steps:
          type: integer
        budget_usd:
          type: number
        time_limit_secs:
          type: integer
        project:
          type: string
        priority:
          type: string
        category:
          type: string
    TaskResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
        tier:
          type: string
        capability_token:
          type: string
        instant_response:
          type: string
    TaskStatusResponse:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
        content:
          type: string
        output:
          type: string
        error:
          type: string
        project:
          type: string
        priority:
          type: string
        category:
          type: string
        created_at:
          type: string
    UpdateTaskRequest:
      type: object
      properties:
        project:
          type: string
        priority:
          type: string
        category:
          type: string
        status:
          type: string
...